+++ slack-gamebot/commands/help.rb
@@ -32,6 +32,7 @@ season: show current season
 Captains
 --------
 set gifs on|off: enable/disable animated GIFs
+set aliases <alias|none> ...: set additional bot aliases
 reset <team>: reset all stats, start a new season
 promote <player>: promote a user to captain
 demote me: demote you from captain